clean up the code

+++ src/main/java/io/mapsmessaging/network/protocol/impl/mqtt_sn/v1_2/packet/PacketFactory.java
@@ -26,7 +26,7 @@ public class PacketFactory {
 
   public MQTT_SNPacket parseFrame(Packet packet) throws IOException {
     int packetLength = packet.get();
-    long dif;
+    int dif;
     if (packetLength == 1) { // Special case so the next 2 bytes indicates the length
       packetLength = MQTTPacket.readShort(packet);
       dif = packet.available() -(packetLength-3);
